+++ examples/pedersen/src/lib.rs
@@ -4,7 +4,7 @@ extern crate alloc;
 use alloc::vec::Vec;
 
 use openzeppelin_crypto::{
-    arithmetic::{uint::U256, BigInteger},
+    arithmetic::uint::U256,
     curve::sw::instance::starknet::StarknetCurveConfig,
     field::prime::PrimeField,
     pedersen::{instance::starknet::StarknetPedersenParams, Pedersen},
@@ -24,14 +24,9 @@ impl PedersenExample {
         let hasher =
             Pedersen::<StarknetPedersenParams, StarknetCurveConfig>::new();
 
-        let inputs: Vec<U256> = inputs
-            .iter()
-            .map(|x| U256::from_bytes_le(&x.to_le_bytes_vec()))
-            .collect();
+        let inputs: Vec<U256> = inputs.iter().map(|x| U256::from(*x)).collect();
 
-        let hash = hasher.hash(inputs[0].into(), inputs[1].into());
-        let hash = hash.expect("Failed to hash").into_bigint().into_bytes_le();
-
-        alloy_primitives::U256::from_le_slice(&hash)
+        let hash = hasher.hash(inputs[0], inputs[1]);
+        hash.expect("Failed to hash").into_bigint().into()
     }
 }